Debug Reporter and finish "main".cpp in laba 1

+++ os_laba_1/Reporter/Reporter.cpp
@@ -6,6 +6,6 @@ int main(int argc, char* argv[]) {
 	char report_name[50];
 	strcpy(binfile_name, argv[1]);
 	strcpy(report_name, argv[2]);
-	int amount = std::atoi(argv[3]);
+	double amount = std::atof(argv[3]);
 	Read_from_bin_file(binfile_name, report_name, amount);
 }
